requirements.txt
pandas
streamlit
matplotlib.pyplot
plotly
nbformat